crypto/octeontx2: fix output field for RSA verify
During RSA sign verification, the OCTEONTX2 PMD returns the decrypted
plaintext in 'sign' field of rte_crypto_rsa_op_param. The 'sign'
field is actually used to pass input to the operation. This PMD
overwrites the 'sign' field buffer. This is non-compliance to lib
cryptodev.

This patch fixes the PMD to use 'cipher' field to return the decrypted
plaintext during RSA verify operation.

Fixes: 04227377c81b ("crypto/octeontx2: support asymmetric in enqueue/dequeue")
Cc: stable@dpdk.org

Signed-off-by: Ramkumar <rbalu@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>