devtools: don't include headline "fix" in backports
It was important in the past to select anything with "fix" in the
headline, but recently more often created false positives and work
to sort tihngs out than identifying many helpful patches.

The community and processes aroudn DPDK matured enough that developers
(rightfully) expect to rely on "Fixes:" and "stable@" marking for
backprots.

Therefore do no more include patches that just include the word fix
in the backport candidate list.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>