net/tap: add a check that Rx/Tx have the same num of queues
Users can create the desired number of RxQ and TxQ in DPDK. For
example, if the number of RxQ = 2 and the number of TxQ = 5,
a total of 8 file descriptors will be created for a tap device,
including RxQ, TxQ, and one for keepalive [1]. The RxQ and TxQ
with the same ID are paired by dup(2) [2].

In this scenario, Kernel will have 3 RxQ where packets are
incoming but not read. The reason for this is that there are only
2 RxQ that are polled by DPDK, while there are 5 queues in the
kernel. This patch add a checking if DPDK has appropriate numbers
of queues to avoid unexpected packet drop.

I had first discussed this issue in OVS [3], but changed my mind
that a fix in DPDK would be more appropriate.

[1]: https://github.com/DPDK/dpdk/blob/0c6e27549c/drivers/net/tap/rte_eth_tap.c#L1967-L1973
[2]: https://github.com/DPDK/dpdk/blob/0c6e27549c/drivers/net/tap/rte_eth_tap.c#L1465-L1466
[3]: https://mail.openvswitch.org/pipermail/ovs-dev/2021-November/389690.html

Signed-off-by: Nobuhiro MIKI <nmiki@yahoo-corp.jp>
Signed-off-by: 0-day Robot <robot@bytheb.org>