eal: allow to exclude memseg from core dump
Some DPDK application is allocated storage partition of 8G(or smaller)
If coredump happens, the application doesn't work because of
insufficient storage space.
The patch provides a config that means whether the memseg memory
is allowed to exclude from core dump.
The DPDK application can choose to open it according to the actual
situation.

Fixes: d72e4042c5eb ("mem: exclude unused memory from core dump")
Cc: stable@dpdk.org

Signed-off-by: Gaoxiang Liu <liugaoxiang@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>