net/spnic: add cmdq and work queue
This commit introduce cmdq and work queue which can be used to
send bulk message data(up to 2KB) to hardware. cmdq provides a
mechanism to encapsulate the message to be sent and handle the
response data or status. work queue is used to manager the wqe
in which includes message data buffer description, ctrl info,
header info and response message data buffer. This patch
implements the initialization and data structure.

Signed-off-by: Yanling Song <songyl@ramaxel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>