net/virtio: fix unreleased resource when creating virtio user dev is failed
When eth_virtio_dev_init is failed, the registered virtio user memory event
cb is not released and the backend created tap device is not destroyed.
It would cause some residual tap device existed in the host and creating
a new vdev could be failed because the new virtio_user_dev could use the
same address pointer and register memory event cb to the same address is
not allowed.

Signed-off-by: Harold Huang <baymaxhuang@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>